refactor(core): refactor ProcessAttributeRepository to separate interface and SQLite implementation
- Create IProcessAttributeRepository interface in domains
- Move SQLite implementation to infrastructure
- Deprecate old ProcessAttributeRepository export for backward compatibility
